Split up tests into separate files to overcome CI failure
